refactor edit_mode.py. moved functions and classes into seperate files.
